please please please expose create_signals
