--- bycycle/kenton_convenience_features/__init__.py
@@ -1,3 +1,0 @@
-"""Burst detection."""
-
-from .simulate_cycles import *
