+++ commit_work.sh
@@ -5,6 +5,8 @@ if [ -z "$1" ]; then
 fi
 
 git add *
+# commit to cluster+recon branch
 git commit -m "$1"
+
 git push
-pip install git+https://github.com/bycycle-tools/bycycle.git@cluster+recon
+pip install -U git+https://github.com/bycycle-tools/bycycle.git@cluster+recon
